Minor tweaks and refactoring
