+++ Loop/Extensions/UIFont.swift
@@ -1,0 +1,30 @@
+//
+//  UIFont.swift
+//  Loop
+//
+//  Created by Michael Pangburn on 2/27/20.
+//  Copyright © 2020 LoopKit Authors. All rights reserved.
+//
+
+import UIKit
+
+
+extension UIFont {
+    convenience init(
+        style: UIFont.TextStyle,
+        design: UIFontDescriptor.SystemDesign,
+        traits: UIFontDescriptor.SymbolicTraits = []
+    ) {
+        var descriptor = UIFontDescriptor.preferredFontDescriptor(withTextStyle: style)
+
+        if #available(iOS 13.0, *),
+            let prettierDescriptor = descriptor.withDesign(design)?.withSymbolicTraits(traits)
+        {
+            descriptor = prettierDescriptor
+        } else if let prettierDescriptor = descriptor.withSymbolicTraits(traits) {
+            descriptor = prettierDescriptor
+        }
+
+        self.init(descriptor: descriptor, size: 0)
+    }
+}
